pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "lightrag-hku>=1.4",
    "mlx>=0.31.2",
    "numpy>=2.0",
    "tokenizers>=0.22.2",
]
